[CCUBE-1510-1][JY] Rename common fields, Remove unnecessary code.
