+++ src/components/element-editor/basic-details/common/index.ts
@@ -1,7 +1,7 @@
-export * from "./description";
-export * from "./id";
-export * from "./label";
+export * from "./description-field";
+export * from "./id-field";
+export * from "./label-field";
 export * from "./mandatory-field";
-export * from "./placeholder";
-export * from "./text-based-preselected-value";
-export * from "./type";
+export * from "./placeholder-field";
+export * from "./text-based-preselected-value-field";
+export * from "./type-field";
